pyproject.toml
[tool.poetry]
name = "rawfinder"
version = "0.3.2"
description = "CLI tool to organize RAW photos by matching JPEG counterparts"
authors = ["Vitalii Shishorin <moskrc@gmail.com>"]
repository = "https://github.com/moskrc/rawfinder"
documentation = "https://moskrc.github.io/rawfinder/"
readme = "README.md"
keywords = ["photography", "raw", "organizer", "cli", "fuji", "canon", "sony", "nikon", "ricoh", "pentax", "olympus", ]
packages = [
    { include = "rawfinder" },
]
include = [
    "LICENSE",
    "CHANGELOG.md"
]

[tool.poetry.dependencies]
python = ">=3.9,<4.0"
rich = "^13.9.4"
pydantic = "^2.10.6"
click = "^8.1.8"
appdirs = "^1.4.4"
pyyaml = "^6.0.1"


[tool.poetry.group.dev.dependencies]
pytest = "^8.2.1"
pytest-cov = "^4.0.0"
pytest-mock = "^3.14.0"
mypy = "^1.11.2"
ruff = "^0.4.4"
pre-commit = "^3.7.1"
types-pyyaml = "^6.0.12.20240330"

[tool.poetry.group.docs.dependencies]
mkdocs = "^1.4.2"
mkdocs-material = "^9.2.7"
mkdocstrings = {extras = ["python"], version = "^0.23.0"}

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.mypy]
files = ["rawfinder"]
disallow_untyped_defs = "True"
disallow_any_unimported = "True"
no_implicit_optional = "True"
check_untyped_defs = "True"
warn_return_any = "True"
warn_unused_ignores = "True"
show_error_codes = "True"

[tool.pytest.ini_options]
testpaths = [
    "tests",
]

[tool.ruff]
target-version = "py39"
line-length = 120
fix = true

[tool.ruff.lint]
select = [
    # flake8-2020
    "YTT",
    # flake8-bandit
    "S",
    # flake8-bugbear
    "B",
    # flake8-builtins
    "A",
    # flake8-comprehensions
    "C4",
    # flake8-debugger
    "T10",
    # flake8-simplify
    "SIM",
    # isort
    "I",
    # mccabe
    "C90",
    # pycodestyle
    "E", "W",
    # pyflakes
    "F",
    # pygrep-hooks
    "PGH",
    # pyupgrade
    "UP",
    # ruff
    "RUF",
    # tryceratops
    "TRY",
]
ignore = [
    # DoNotAssignLambda
    "E731",
]

[tool.ruff.format]
preview = true

[tool.coverage.report]
skip_empty = true

[tool.coverage.run]
branch = true
source = ["rawfinder"]


[tool.ruff.per-file-ignores]
"tests/*" = ["S101"]

[tool.poetry.scripts]
rawfinder = "rawfinder.cli:cli"
